Replaced `faceup` tests.
The `faceup` framework isn't suitable for this kind of testing since
between different versions of Emacs the major mode font-locking can
change.

Rather than having multiple different .faceup files for different
versions of Emacs, it's better to just verify that the expected faces
are present at a fixed set of positions in a buffer with the tested
major modes.